backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
pydantic==2.0.3
crewai==0.28.1
langchain==0.0.335
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
httpx==0.25.0 